requirements.txt
Pillow==9.0.1
Pillow==10.3.0
Requests==2.32.3
streamlit==1.35.0
streamlit_lottie==0.0.5
streamlit_option_menu==0.3.13
